created tables with JPA, added relationships between tables and updated our classes and attributes

+++ src/main/java/be/kdg/programming3/repository/ObstacleRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ObstacleRepository extends JpaRepository<Obstacle, Integer> {
+public interface ObstacleRepository extends JpaRepository<Obstacle, Long> {
 
 }
